refactor: 'url'필드명을 'thumbnailUrl'로 변경
